fix create, edit repos functions

+++ internal/model/freelancer.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type Freelancer struct {
-	ID                int       `json:"id"`
+	ID                int64       `json:"id"`
 	AccountId         int64       `json:"accountId"`
 	RegistrationDate  time.Time `json:"registrationDate"`
 	Country           string    `json:"country"`
@@ -12,6 +12,6 @@ type Freelancer struct {
 	Phone             string    `json:"phone"`
 	TagLine           string    `json:"tagline"`
 	Overview          string    `json:"overview"`
-	ExperienceLevelId int       `json:"experienceLevelId"`
-	SpecialityId      int       `json:"specialityId,string"`
+	ExperienceLevelId int64       `json:"experienceLevelId"`
+	SpecialityId      int64       `json:"specialityId,string"`
 }
