+++ internal/model/hireManager.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type HireManager struct {
-	ID					int		`json:"id"`
+	ID					int64		`json:"id"`
 	AccountID 			int64		`json:"accountId"`
 	RegistrationDate	time.Time	`json:"registrationDate"`
 	Location			string 		`json:"location"`
-	CompanyID			int 		`json:"companyId"`
+	CompanyID			int64 		`json:"companyId"`
 }
