+++ internal/model/notification.go
@@ -1,7 +1,7 @@
 package model
 
 type Notification struct {
-	UserID int `json:"-"`
+	UserID int64 `json:"-"`
 	NewMessages bool `json:"new_messages"`
 	NewProjects bool `json:"new_projects"`
 	NewsFromService bool `json:"news_service"`
